ps_disk_usage() for Linux
The macOS implementation is actually POSIX,
so we can just use that.
